+++ teste/teste.xcworkspace/contents.xcworkspacedata
@@ -2,7 +2,7 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "group:teste.xcodeproj">
+      location = "group:/Users/rmarangoni/Documents/Diversos/challenge-ios/teste/lodjaApp.xcodeproj">
    </FileRef>
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
