--- teste/teste/ProductViewController.swift
@@ -1,27 +1,0 @@
-//
-//  ProductViewController.swift
-//  teste
-//
-//  Created by Rodrigo Marangoni on 29/01/18.
-//  Copyright © 2018 romarangoni. All rights reserved.
-//
-
-import UIKit
-
-class ProductViewController: UIViewController {
-    @IBOutlet var productView: ProductView!
-    var product: Product!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        self.productView.configure(self.product)
-        self.title = self.product.nome
-
-    }
-    @IBAction func doReserve(_ sender: Any) {
-        
-    }
-    
-
-
-}
